Simplify security configuration
