+++ ngrinder-controller/src/main/java/org/ngrinder/infra/config/SvnSecurityConfig.java
@@ -1,0 +1,156 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ngrinder.infra.config;
+
+import static org.ngrinder.common.constant.ControllerConstants.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.ngrinder.security.NGrinderAuthenticationPreAuthProvider;
+import org.ngrinder.security.NGrinderAuthenticationProvider;
+import org.ngrinder.security.NGrinderUserDetailsService;
+import org.ngrinder.security.SvnHttpBasicEntryPoint;
+import org.ngrinder.security.UserSwitchPermissionVoter;
+import org.ngrinder.user.service.UserService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.access.AccessDecisionVoter;
+import org.springframework.security.access.vote.*;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsByNameServiceWrapper;
+import org.springframework.security.crypto.password.ShaPasswordEncoder;
+import org.springframework.security.web.access.expression.DefaultWebSecurityExpressionHandler;
+import org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationToken;
+
+import lombok.AllArgsConstructor;
+
+/**
+ * Some User want to have more secured password. Provide the enhanced pw with sha256 if a user
+ * specifies ngrinder.security.sha256 in system.conf
+ */
+@Configuration
+@EnableWebSecurity
+@Order(1)
+@AllArgsConstructor
+public class SvnSecurityConfig extends WebSecurityConfigurerAdapter {
+
+	private Config config;
+
+	private NGrinderUserDetailsService ngrinderUserDetailsService;
+
+	private UserSwitchPermissionVoter userSwitchPermissionVoter;
+
+	private SvnHttpBasicEntryPoint svnHttpBasicEntryPoint;
+
+	private NGrinderAuthenticationProvider nGrinderAuthenticationProvider;
+
+	private UserService userService;
+
+	/**
+	 * Provide the appropriate shaPasswordEncoder depending on the ngrinder.security.sha256 config.
+	 *
+	 * @return ShaPasswordEncoder with "SHA-256" algorithm if ngrinder.security.sha256=true. Otherwise
+	 *         returns with "SHA-1"
+	 */
+	@SuppressWarnings("deprecation")
+	@Bean
+	public ShaPasswordEncoder shaPasswordEncoder() {
+		boolean useEnhancedEncoding = config.getControllerProperties().getPropertyBoolean(PROP_CONTROLLER_USER_PASSWORD_SHA256);
+		return useEnhancedEncoding ? new ShaPasswordEncoder("SHA-256") : new ShaPasswordEncoder("SHA-1");
+	}
+
+	/**
+	 * Generic Web
+	 * @return AuthenticatedVoter
+	 */
+	@Bean
+	public AuthenticatedVoter authenticatedVoter() {
+		return new AuthenticatedVoter();
+	}
+
+	@Bean
+	public DefaultWebSecurityExpressionHandler webSecurityExpressionHandler() {
+		DefaultWebSecurityExpressionHandler defaultWebSecurityExpressionHandler = new DefaultWebSecurityExpressionHandler();
+		defaultWebSecurityExpressionHandler.setDefaultRolePrefix("");
+		return defaultWebSecurityExpressionHandler;
+	}
+
+	@Bean
+	public RoleVoter roleVoter() {
+		RoleVoter roleVoter = new RoleVoter();
+		roleVoter.setRolePrefix("");
+		return roleVoter;
+	}
+
+	/**
+	 * Svn AccessDecisionManager
+	 * @return UnanimousBased
+	 */
+	@Bean
+	public UnanimousBased svnAccessDecisionManager() {
+		List<AccessDecisionVoter<?>> decisionVoters = new ArrayList<>();
+		decisionVoters.add(authenticatedVoter());
+		decisionVoters.add(userSwitchPermissionVoter);
+		return new UnanimousBased(decisionVoters);
+	}
+
+	@Override
+	protected void configure(HttpSecurity http) throws Exception {
+		//SVN
+		http
+			.requestMatchers().antMatchers("/svn/**")
+			.and()
+			.authorizeRequests()
+			.antMatchers("/svn/**")
+			.hasAnyRole("A", "S", "U")
+			.accessDecisionManager(svnAccessDecisionManager())
+			.and()
+			.httpBasic().authenticationEntryPoint(svnHttpBasicEntryPoint)
+			.realmName("svn")
+			.and()
+			.csrf().disable();
+	}
+
+	@Override
+	public void configure(WebSecurity web) {
+		web.expressionHandler(webSecurityExpressionHandler());
+	}
+
+	@Override
+	protected void configure(AuthenticationManagerBuilder auth) {
+		auth.authenticationProvider(ngrinderPreAuthProvider()).authenticationProvider(nGrinderAuthenticationProvider);
+	}
+
+	@Bean
+	public NGrinderAuthenticationPreAuthProvider ngrinderPreAuthProvider() {
+		NGrinderAuthenticationPreAuthProvider nGrinderAuthenticationPreAuthProvider = new NGrinderAuthenticationPreAuthProvider();
+		nGrinderAuthenticationPreAuthProvider.setPreAuthenticatedUserDetailsService(userDetailsServiceWrapper());
+		nGrinderAuthenticationPreAuthProvider.setUserService(userService);
+		return nGrinderAuthenticationPreAuthProvider;
+	}
+
+	@Bean
+	public UserDetailsByNameServiceWrapper<PreAuthenticatedAuthenticationToken> userDetailsServiceWrapper() {
+		UserDetailsByNameServiceWrapper<org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationToken> userDetailsByNameServiceWrapper = new UserDetailsByNameServiceWrapper<>();
+		userDetailsByNameServiceWrapper.setUserDetailsService(ngrinderUserDetailsService);
+		return userDetailsByNameServiceWrapper;
+	}
+
+}
